update: Internal Activity Recipient Wording
This updates the wording for the Internal Activity Notice's Assigned
Agent/Team recipient. It used to use the terminology "or" to say the
alert is sent to an Agent -or- Team. However, the code was updated to
send the alert to everyone assigned so the docs needed to be updated to
reflect this. This simply changes the word "or" to "and" to be more
clear about who will receive the alert.

Initial Report:
- https://github.com/osTicket/osTicket/issues/6659