update: MS OAuth2 Outlook API Deprecations 2024
This updates the documentation to add notices about the recent MS
OAuth2 Outlook endpoint deprecations. This adds information about what
happened, what the resulting error was, and provides steps on how to
resolve the issue. In addition this updates all OAuth2 docs to provide a
notice about having the OAuth2 Plugin installed and enabled before
proceeding with the setup with a link to the relevant docs. Lastly, this
updates the Copyright.